backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
distlib==0.4.0
django==6.0.3
django-environ==0.13.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
filelock==3.25.2
identify==2.6.18
idna==3.11
kombu==5.6.2
lxml==6.0.2
nodeenv==1.10.0
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pre-commit==4.5.1
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pyjwt==2.12.1
python-dateutil==2.9.0.post0
python-discovery==1.2.0
python-dotenv==1.2.2
pyyaml==6.0.3
redis==7.3.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing-extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
virtualenv==21.2.0
wcwidth==0.6.0
pgvector==0.3.6
